smarthome 2.0 openWB 1.9 Zerlegung
Der Smarthomemq.py wurde zerlegt in einen openWB 1.9 Teil und einen zwischen openWB 1.9 und openWB 2.0 gemeinsamen Teil. Die gleiche Einlieferung findet auch für openWB 2.0 statt. Somit soll dieser PR nur für openWB 1.9 verwendet werden Es wurde ein Fehler beim mindestlaufzeit pro Tag behoben
